+++ web_app/app.py
@@ -1,29 +1,23 @@
-import os
-import sys
-import mediapipe as mp
-import tensorflow as tf
 from os.path import dirname, abspath, join
-# Get the directory of the current file (app.py) and find the parent directory
+import sys
 current_dir = dirname(abspath(__file__))
 parent_dir = dirname(current_dir)
-
-# Add the 'src' directory to the path
 sys.path.append(join(parent_dir, 'src'))
-import CNN_model as cnn
 import loading_dataset as ld
 sys.path.append(current_dir)
+import os
+import mediapipe as mp
 from flask import Flask, render_template, request
 import numpy as np
-from torch.utils.data import Dataset,DataLoader
 from torch import tensor, float32, argmax
 from datetime import datetime
-import torchmetrics
 from glob import glob
 from PIL import Image
 import predict
 
 DEVICE = ld.load_device()
 
+# get most recent image and shaped in right size and format
 def croped_shaped_image(folder_path):
     if not os.path.isdir(folder_path):
         print(f"Error: Folder '{folder_path}' does not exist.")
@@ -56,26 +50,18 @@ def getLetter():
     X_array =X_array.astype(np.float32)
     mlp = predict.load_mlp_model()
     prediction = mlp(tensor(X_array.reshape(1,21,2),dtype=float32).to(DEVICE)) # reshape to incorporate batch size
-    letter_prediction = argmax(prediction) #letter is of highest probability
-    # for x in loader:
-    #     print(x)
-    #     x = x.to(DEVICE)
-    #     letter_prediction = mlp(x) # error here
+    letter_prediction = argmax(prediction) # letter is of highest probability
     return predict.ALPHABET[letter_prediction]
 
 def save_image(image_array, folder_path, filename):
   if not os.path.exists(folder_path):
     os.makedirs(folder_path)
-
-  # Convert the NumPy array to a PIL Image object (assuming RGB format)
+  # Convert the NumPy array to a PIL Image
   image = Image.fromarray(image_array.astype(np.uint8))
-
   # Save the image as a PNG file
   image.save(os.path.join(folder_path, filename))
-
   print(f"Image saved successfully: {os.path.join(folder_path, filename)}")
   
-
 app = Flask(__name__)
 
 @app.route("/") # rendering html page
@@ -85,9 +71,9 @@ def index():
 @app.route("/receive",methods=["POST"])
 def receive_data():
     data = request.get_json()
+    # image array from main.jss
     image_data = data.get('imageData', [])
     image_data = np.asarray(image_data)
-
     image_data = image_data[:, :, :3]# (480, 640, 3)
     timestamp = datetime.now().strftime("%Y%m%d-%H%M%S")
     directory_path = 'images/saved_arrays'
@@ -95,7 +81,6 @@ def receive_data():
     filename = f"image_data_{timestamp}.png"
     save_image(image_data, directory_path, filename)
     letter = getLetter()
-    
     return {'letter': letter}
 
 if __name__ == "__main__":
