+++ web_app/templates/index.html
@@ -4,14 +4,14 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>MAIS-202-F2023-ASL-PROCESSING</title>
-    <link rel="icon" href="/images/logo-icon.png">
+    <!--<link rel="icon" href="../images/logo.png"> -->
     <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
     <script src="{{ url_for('static', filename='main.js') }}" defer></script>   
 </head>
 <body>
     <header>
         <div id="top">
-            <img src="/images/logo.png" alt="logo" width="175" height="150" class="logo">
+            <!--<img src="../images/logo.png" alt="logo" width="175" height="150" class="logo">-->
             <h1>American Sign Language Translator</h1>
             <br style="height:20px">
         </div>
@@ -38,6 +38,7 @@
         <!--Buttons: Start, Snap and Stop-->
         <button id="startButton">Start Webcam</button>
         <button id="snapButton" disabled>Take Picture</button>
+        <button id="deleteButton" disabled>Delete</button>
         <button id="stopButton" disabled>Stop Webcam</button>
         <!--Displayed Image generated by the Snap Button used to save Image-->
         <div id="images">
